Add custom handler functionality
This adds a new annotation `Handler` that can be used
to tie a config member to a custom handler function
that is called when the value is set via a conf file
or through the CLI.

The handler is then responsible to set the config
member.

This is implemented by creating a wrapper function,
called proxy, that imports the handler function and
calls it with the config struct member as context.